Fix single port RRQ with options fail
When running tftpd with -s single port argument and Read Request with options
are provided ACK for OACK is not properly received.

The problem is that code is waiting MPSC channel response, but there is no
thread for receiving data from UDP socket and MPSC receive will timeout.

Fix is to move waiting of ACK response for OACK into worker thread in the
beginning of file sending. I.e. File send is started with receiving ACK when
there were options and OACK was sent in the accept_request().

Fixes issue #16